requirements.txt
azure-openai>=1.0.0
openai>=1.6.0
streamlit>=1.28.0
databricks-sdk>=0.20.0
databricks-connect>=14.0.0
pytest>=7.4.0
pytest-mock>=3.11.1
langchain>=0.1.0
python-dotenv>=1.0.0
requests>=2.31.0
opentelemetry-api>=1.19.0
opentelemetry-sdk>=1.19.0
opentelemetry-exporter-otlp>=1.19.0
opentelemetry-instrumentation>=0.40b0
pydantic>=2.0.0
agentssdk>=0.5.0 